animeNightScript/requirements.txt
## The following requirements were added by pip freeze:
astroid==2.4.2
colorama==0.4.3
feedparser==5.2.1
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.6.3
mccabe==0.6.1
pathlib==1.0.1
pylint==2.5.3
python-dotenv==0.15.0
six==1.15.0
toml==0.10.1
wrapt==1.12.1